refactor(config): convert manage config to TypeScript and update related imports
